moving and renaming HttStylesNew
