+++ interface/HttStylesNew.h
